requirements.txt
PyQt6>=6.5.0
python-docx>=1.1.0
